pyproject.toml
[tool.poetry]
name = "ragbot"
version = "0.1.0"
description = "A RAG-powered tool to enhance chatbots responses."
authors = ["Rubén Martínez Amodia <ruben.martinezam@alumnos.unican.es>"]
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.11,<4.0"
langchain = "^0.3.25"
langsmith = "^0.1.142"
ipython = "^9.2.0"
sacrebleu = "^2.5.1"
rouge-score = "^0.1.2"
chromadb = "^1.0.8"
langchain-community = "^0.3.23"
google-generativeai = "^0.8.5"
langchain-google-genai = "^2.0.4"
langchain-huggingface = "^0.2.0"
langchain-ollama = "^0.3.2"


[tool.poetry.group.dev.dependencies]
black = "^24.3"
isort = "^6.0.1"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.1"
mkdocs-material = "^9.6.12"
mkdocstrings = {extras = ["python"], version = "^0.29.1"}


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"